requirements.txt
keras
opencv-python
opencv-contrib-python
numpy<2
schedule
requests
PyYAML
ultralytics
torch
moviepy
pandas
scikit-learn